Renaming code to MetOceanViewer from Adcirc Validator.
