+++ src/timeseries_add_data.cpp
@@ -35,7 +35,7 @@
 
 #include <timeseries_add_data.h>
 #include <ui_timeseries_add_data.h>
-#include <ADCvalidator.h>
+#include <MetOceanViewer.h>
 
 int NumIMEDSFiles = 0;
 int CurrentRowsInTable = 0;
