+++ ui/MetOceanViewer_main.ui
@@ -23,7 +23,7 @@
    </size>
   </property>
   <property name="windowTitle">
-   <string>ADCIRC Validator</string>
+   <string>MetOceanViewer</string>
   </property>
   <property name="windowIcon">
    <iconset resource="../ADCvalidatior.qrc">
@@ -3352,7 +3352,7 @@
      <x>0</x>
      <y>0</y>
      <width>780</width>
-     <height>19</height>
+     <height>21</height>
     </rect>
    </property>
    <widget class="QMenu" name="menuFile">
